backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
aiohttp
tldextract
sentence-transformers
groq
Pillow
imagehash
python-whois
dnspython
requests
email-validator
beautifulsoup4
